results/rebuttal_ablation_no_summarization/run-20231115-165815_LLMatic_donnemartin-system-design-oop-whatsapp_0_20-runs_log/whatsapp/8/LLMatic/connectivity.py
class Connectivity:
	def __init__(self):
		self.online = False
		self.message_queue = []

	def check_connectivity(self):
		# This is a mock method. In a real-world application, this method would check the actual connectivity of the system.
		return self.online

	def queue_message(self, message):
		self.message_queue.append(message)

	def display_status(self):
		return 'Online' if self.online else 'Offline'

	def connect(self):
		self.online = True
		# Send all queued messages
		for message in self.message_queue:
			# In a real-world application, this would send the message
			print(f'Sending message: {message}')
		self.message_queue = []

	def disconnect(self):
		self.online = False
